requirements.txt
# 核心依赖
torch>=2.0.0
transformers>=4.40.0
gradio>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0

# 可选依赖（用于加速和优化）
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# 工具依赖
python-dotenv>=1.0.0
pyyaml>=6.0